refatoring: move - ClientRegisterController (controller) -> (Controller/Register)
